Enable label on ellipse annotation
